Remove test_pandas_interop.py
